+++ src/main/java/roomescape/controller/ClientRankController.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.dto.ThemeResponse;
+import roomescape.dto.response.ThemeResponse;
 import roomescape.service.RankService;
 
 import java.util.List;
