+++ src/main/java/roomescape/controller/ClientReservationController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.dto.BookResponse;
+import roomescape.dto.response.BookResponse;
 import roomescape.service.BookService;
 
 import java.time.LocalDate;
