+++ src/main/java/roomescape/dto/response/TimeSlotResponse.java
@@ -1,4 +1,4 @@
-package roomescape.domain.dto;
+package roomescape.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import roomescape.domain.TimeSlot;
