+++ src/main/java/roomescape/repository/ThemeDao.java
@@ -7,7 +7,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Theme;
-import roomescape.domain.dto.ThemeRequest;
+import roomescape.dto.request.ThemeRequest;
 
 import java.util.List;
