+++ src/test/java/roomescape/controller/ClientRankTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.domain.dto.ThemeResponse;
+import roomescape.dto.response.ThemeResponse;
 
 import java.util.List;
