+++ src/test/java/roomescape/controller/ClientReservationTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.domain.dto.TimeSlotResponse;
+import roomescape.dto.response.TimeSlotResponse;
 
 import java.util.List;
